+++ packages/core/src/index.ts
@@ -2,5 +2,6 @@ export { findMatchingImage } from './match.ts';
 export * from './env.ts';
 export * from './config.ts';
 export * from './debug.ts';
+export * from './lqip.ts';
 
 export type * from './types.ts';
