+++ packages/core/src/lqip.ts
@@ -1,0 +1,5 @@
+import { Lqip, LqipBlurhash } from './types';
+
+export function isLqipBlurhash(lqip: Lqip | undefined): lqip is LqipBlurhash {
+  return lqip?.type === 'blurhash';
+}
